0.2.1 dodano mozliwosc dodawania nowego produktu

+++ FitApp/ObliczCele.cs
@@ -43,7 +43,7 @@ namespace FitApp
             }
             else
             {
-                return (int)zapotrzebowanie;
+                return zapotrzebowanie;
             }
         }
